Fix demo breaking changes
