+++ runtimes/deno/versions/1.21/src/fileStreamReady.ts
@@ -1,0 +1,3 @@
+export async function fileStreamReady(file: Deno.FsFile) {
+  return;
+}
