+++ runtimes/deno/versions/1.21/src/getContextBody.ts
@@ -1,0 +1,6 @@
+export async function getContextBody(ctx: any) {
+  return await ctx.request.body({
+    type: "bytes",
+    limit: 20 * 1024 * 1024,
+  }).value;
+}
