+++ runtimes/deno/versions/1.24/src/app.ts
@@ -1,0 +1,5 @@
+import { Application } from "https://deno.land/x/oak@v10.6.0/mod.ts";
+
+export function createApp() {
+  return new Application();
+}
