+++ runtimes/deno/versions/1.35/src/app.ts
@@ -1,0 +1,5 @@
+import { Application } from "https://deno.land/x/oak@v14.1.1/mod.ts";
+
+export function createApp() {
+  return new Application();
+}
