+++ runtimes/deno/versions/latest/src/app.ts
@@ -1,0 +1,5 @@
+import { Application } from "https://deno.land/x/oak@v17.0.0/mod.ts";
+
+export function createApp() {
+  return new Application();
+}
