+++ runtimes/deno/versions/latest/src/fileStreamReady.ts
@@ -1,0 +1,4 @@
+export async function fileStreamReady(file: Deno.FsFile) {
+  await file.ready;
+  return;
+}
