+++ runtimes/deno/versions/latest/src/getContextBody.ts
@@ -1,0 +1,3 @@
+export async function getContextBody(ctx: any) {
+  return await ctx.request.body.arrayBuffer();
+}
