+++ tests/resources/functions/deno/1.21/deps.ts
@@ -1,0 +1,4 @@
+export {
+  decode as decodeBase64,
+  encode as encodeBase64,
+} from "https://deno.land/std@0.202.0/encoding/base64.ts";
