+++ tests/resources/functions/deno/1.35/deps.ts
@@ -1,0 +1,4 @@
+export {
+  decodeBase64,
+  encodeBase64,
+} from "https://deno.land/std@0.224.0/encoding/base64.ts";
