+++ tests/resources/functions/deno/latest/deps.ts
@@ -1,0 +1,1 @@
+export { decodeBase64, encodeBase64 } from "jsr:@std/encoding";
